requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-wtf==1.2.2
pillow==11.1.0
opencv-python==4.11.0.86
numpy==2.1.3
tensorflow==2.19.0
werkzeug==3.1.3
jinja2==3.1.6
wtforms==3.2.1
email-validator==2.1.0 